# bpo-47070: improve performance of array_inplace_repeat