--- a/Objects/bytearrayobject.c
+++ b/Objects/bytearrayobject.c
@@ -339,11 +339,9 @@ bytearray_irepeat(PyByteArrayObject *self, Py_ssize_t count)
 {
     if (count < 0)
         count = 0;
-    else {
-        if (count == 1) {
+    else if (count == 1) {
             Py_INCREF(self);
             return (PyObject*)self;
-        }
     }
 
     const Py_ssize_t mysize = Py_SIZE(self);