The `array_inplace_repeat` is inefficient for small arrays and a high number of repeats. This can be improved by using the same approach as in the `array_inplace` method (same approach as https://bugs.python.org/issue47005)

Microbenchmark:
```
import pyperf
runner = pyperf.Runner()

setup="import array; a=array.array('d',[1,2,3,])"
code= "a*=1"
runner.timeit('repeat 1', stmt=code, setup=setup)
setup="import array;"
code= "a=array.array('d',[1,2,3,]); a*=1000"
runner.timeit('repeat 1000', stmt=code, setup=setup)              
```
Results in:
```
repeat 1: Mean +- std dev: [base] 53.0 ns +- 1.4 ns -> [patch] 40.4 ns +- 0.7 ns: 1.31x faster
repeat 1000: Mean +- std dev: [base] 2.78 us +- 0.08 us -> [patch] 908 ns +- 52 ns: 3.06x faster

Geometric mean: 2.01x faster
```

<!--
Thanks for your contribution!
Please read this comment in its entirety. It's quite important.

# Pull Request title

It should be in the following format:

```
bpo-NNNN: Summary of the changes made
```

Where: bpo-NNNN refers to the issue number in the https://bugs.python.org.

Most PRs will require an issue number. Trivial changes, like fixing a typo, do not need an issue.

# Backport Pull Request title

If this is a backport PR (PR made against branches other than `main`),
please ensure that the PR title is in the following format:

```
[X.Y] <title from the original PR> (GH-NNNN)
```

Where: [X.Y] is the branch name, e.g. [3.6].

GH-NNNN refers to the PR number from `main`.

-->