@eendebakpt 2022-03-20
@sweeneyde Based on your suggestion I created utility methods `_PyBytes_Repeat` and `_PyBytes_RepeatInPlace`. They are in `bytearrayobject.h`, but perhaps there is a better location.

I defined the objects to be `static inline`, but am not sure they are inline on all platforms. If not, there could be (small) performance difference for the `bytes` and `bytearray` objects due to the overhead of calling the utility methods.