@sweeneyde 2022-03-20
They should go in Include/internal/pycore_bytes.h so they don't leak into the public API.

I wouldn't worry too much about whether things will get inlined -- if the compiler decides it's beneficial, it can inline it. Also, I'm guessing the single function call overhead is vastly outscaled by the several memcpy calls.